Fixing a network problem that prevents skyDNS pod from starting
I ran across a similar problem to the one I was having in getting
skyDNS to run properly. It was at:
https://github.com/kubernetes/kubernetes/issues/34101
The problem was that only 1 of the three containers in the pod would
start up correctly. After readding this thread there were two things
that I hadn't done in the solution. This first was to start kube-proxy
with --proxy-mode=userspace rather than iptables. I made this change
and retry starting skyDNS. It comes fully up now. Still need to test
that it is working correctly.